+++ test/doing_import_test.rb
@@ -9,6 +9,7 @@ require 'test_helper'
 # Tests for entry modifying commands
 class DoingImportTest < Test::Unit::TestCase
   include DoingHelpers
+  ENTRY_REGEX = /^\d{4}-\d\d-\d\d \d\d:\d\d \|/.freeze
   ENTRY_TS_REGEX = /\s*(?<ts>[^|]+) \s*\|/.freeze
   ENTRY_DONE_REGEX = /@done\((?<ts>.*?)\)/.freeze
 
@@ -56,8 +57,9 @@ class DoingImportTest < Test::Unit::TestCase
   end
 
   def assert_count_entries(count, shown, message = 'Should be X entries shown')
-    assert_equal(count, shown.uncolor.strip.split("\n").count, message)
+    assert_equal(count, shown.uncolor.strip.scan(ENTRY_REGEX).count, message)
   end
+
 
   def mktmpdir
     tmpdir = Dir.mktmpdir