+++ test/doing_output_test.rb
@@ -8,6 +8,7 @@ require 'test_helper'
 # Tests for entry modifying commands
 class DoingOutputTest < Test::Unit::TestCase
   include DoingHelpers
+  ENTRY_REGEX = /^\d{4}-\d\d-\d\d \d\d:\d\d \|/.freeze
 
   def setup
     @tmpdirs = []
@@ -48,6 +49,25 @@ class DoingOutputTest < Test::Unit::TestCase
     assert_match(/#{subject}\s*$/, result, 'doing show @tag1 results should include test entry')
   end
 
+  def test_show_command_tag_boolean
+    subject = 'Test new entry @tag1'
+    doing('now', subject)
+    subject2 = 'Test new entry 2 @tag2'
+    doing('now', subject2)
+    subject3 = 'Test new entry 3 @tag1 @tag2 @tag3'
+    doing('now', subject3)
+
+    result = doing('show', '--tag', 'tag1,tag2', '--bool', 'and').uncolor.strip
+    assert_count_entries(1, result, 'There should be 1 entry shown with both @tag1 and @tag2')
+    assert_match(/#{subject3}\s*$/, result, 'doing show results should include entry with both @tag1 and @tag2')
+
+    result = doing('show', '--tag', 'tag1,tag2', '--bool', 'or').uncolor.strip
+    assert_count_entries(3, result, 'There should be 3 entries shown with either @tag1 or @tag2')
+    result = doing('show', '--tag', 'tag2', '--bool', 'not').uncolor.strip
+    assert_count_entries(1, result, 'There should be 1 entry shown without @tag2')
+    assert_match(/#{subject}\s*$/, result, 'doing show results should include entry without @tag2')
+  end
+
   def test_today_command
     subject = 'Test new entry @tag1'
     doing('done', subject)
@@ -76,13 +96,13 @@ class DoingOutputTest < Test::Unit::TestCase
     doing('now', 'Test new entry @tag1')
     doing('now', 'Test new entry 2 @tag2')
     result = doing('--stdout', 'on', 'today')
-    assert_count_entries(3, result, 'There should be 2 entries and a date interpreted line')
+    assert_count_entries(2, result, 'There should be 2 entries')
   end
 
   private
 
-  def assert_count_entries(count, shown, message = "Should be X matching entries")
-    assert_equal(count, shown.uncolor.strip.split("\n").count, message)
+  def assert_count_entries(count, shown, message = 'Should be X entries shown')
+    assert_equal(count, shown.uncolor.strip.scan(ENTRY_REGEX).count, message)
   end
 
   def mktmpdir